home20ago20

+++ server/sistema-condominio/src/main/java/com/siscondominio/security/UserSS.java
@@ -13,7 +13,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 public class UserSS implements UserDetails {
 	private static final long serialVersionUID = 1L;
 	
-	private Long id;
+	private Integer id;
 	private String email;
 	private String senha;
 	private Collection<? extends GrantedAuthority> authorities;
@@ -21,7 +21,7 @@ public class UserSS implements UserDetails {
 	public UserSS() {
 	}
 	
-	public UserSS(Long id, String email, String senha, Set<Perfil> perfis) {
+	public UserSS(Integer id, String email, String senha, Set<Perfil> perfis) {
 		super();
 		this.id = id;
 		this.email = email;
@@ -29,7 +29,7 @@ public class UserSS implements UserDetails {
 		this.authorities = perfis.stream().map(x -> new SimpleGrantedAuthority(x.getDescricao())).collect(Collectors.toList());
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
